Refactor components and update icons; replace DollarSign with IndianRupee, reorganize imports, and remove unused Footer component.
